backend/requirements.txt
Flask
Flask-CORS
google-generativeai
python-dotenv
Flask-SQLAlchemy
pandas